refactor(senpi-skills): extract Senpi compatibility guidance injection
sync-skills.mjs was 294 pure LOC after the native skill source extraction,
still above the 250-LOC repo ceiling. Extract the Senpi harness
tool-compatibility banner injection responsibility (opencodeOnlyOrchestration
pattern, senpiHarnessToolCompatibility banner string, end markers, and the
four functions that detect/remove/insert the banner) into a focused
senpi-compatibility-guidance.mjs module.

The orchestrator imports insertSenpiCompatibilityGuidance and delegates to
it inside applySharedTierAdaptation. sync-skills.mjs drops to 238 pure LOC;
the new module is 65 pure LOC; the existing native-skill-sources.mjs remains
at 55. All three are well under the 250-LOC ceiling.

A focused test (senpi-compatibility-guidance.test.mjs) proves banner
insertion, no-op on clean content, idempotence with pre-existing banners,
stale-banner replacement, and the canonical end marker.

Plan: .omo/plans/senpi-onboarding-and-init-deep-advisor.md
